Fixed various bugs found while testing the db changes

+++ objects/database_models/animote_user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, BigInteger
 
 from objects.database_models import Base
 
@@ -6,7 +6,7 @@ from objects.database_models import Base
 class AnimoteUser(Base):
     __tablename__ = "animote_users"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
 
     def __repr__(self):
         return "<AnimoteUser(id={0.id})>".format(self)
